Added RDP to Wireshark
Explanation:
Xvfb was added to create a virtual screen.
xrdp was added to allow RDP connections.
tigervnc was added to listen on localhost and
allow xrdp to connect to Xvfb.

The startup command was then moved into a
script to allow all three services to start.

Reason:
This modification allows Wireshark to be run
remotely on a headless server and with reduced
privileges.

Example cases:
docker run -it --rm --network=host \
                    --cap-add net_raw \
                    --cap-add net_admin \
                    --name=wireshark \
                    jaredharringtongibbs/wireshark

Tags:
[remote]
[wireshark]