Add bigchaindb start script and client
Explanation:
I added a script to autolaunch a bigchaindb instance.
I also added a client with all the requirements to
connect to the bigchaindb instance.

Reason:
It automates the creation of a bigchiandb installation
with little input require from the user and thus
makes it easier to try bigchaindb.

Example cases:
null

Tags:
[bigchaindb]
[addition]