Created docker to demonstrate code "hiding"
Explanation:
This docker was created to test/demonstrate
that an automated build could be generated and
a image could be pushed to overwrite the actual
docker image hosted on hub.docker.com.

Reason:
This makes it easy for students to grasp the
importance of inspecting images before running
them.

Example cases:
null

Tags:
[bad-docker]